// frontbox-pin-console/web/src/types/generated/ResolvedExpansionBoard.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FastExpansionBoardModels } from "./FastExpansionBoardModels";
import type { ResolvedLedPort } from "./ResolvedLedPort";

export type ResolvedExpansionBoard = { address: number, breakout: number | null, led_ports: Array<ResolvedLedPort>, model: FastExpansionBoardModels, };